### 2021-07-14 #452 @vgherard
Pre-submission inquiry for {kgrams}: Classical k-gram Language Models
Submitting Author: Valerio Gherardi (@vgherard)
Repository:  https://github.com/vgherard/kgrams
Submission type: Pre-submission

---

```
Package: kgrams
Title: Classical k-gram Language Models
Version: 0.1.0.9000
    person(given = "Valerio",
           family = "Gherardi",
           email = "vgherard@sissa.it",
        Tools for training and evaluating k-gram language models in R,
        supporting several probability smoothing techniques,
        perplexity computations, random text generation and more.
SystemRequirements: C++11
    Rcpp, RcppProgress
    Rcpp, rlang, methods, utils,  RcppProgress (>= 0.1), Rdpack
    covr,
RdMacros: Rdpack
    https://github.com/vgherard/kgrams

```

	- [ ] text data

This package implements classical k-gram language model algorithms, including utilities for training, evaluation and text prediction. Language models are an angular stone of Natural Language Processing applications, and the conceptual simplicity of k-gram models makes them a good model baseline, also of pedagogical value.

k-gram models are a simple form of Machine-Learning applied to text data; as such, machine-learning is definitely the most appropriate category within the above ones. I would be inclined to define this as an "Unsupervised" learning problem, since the target function being learned (the language's probability distribution over sentences) is clearly not explicit in the training data - but have never seen this particular qualification in the literature.

Not yet (NB: this is a presubmission inquiry).

The package can be useful for students and/or researchers, for performing small-scale experiments with Natural Language Processing. In addition, it might be helpful in the building of more complex language models, for quick baseline modeling.

I am not aware of any R package with same purpose and functionalities of `kgrams`. The CRAN package [ngram](https://cran.r-project.org/web/packages/ngram/index.html) has some relative overlap in scope, in that it provides k-gram tokenization algorithms and random text generation, but offers no support for language model algorithms.

1. The package was accepted some months ago by CRAN.
1. Despite the "lifecycle:experimental" badge and the development version number, I haven't made important API changes or addied features to this package for a long time. With this submission, I'm hoping to receive your feedback regarding possible improvements and/or whether the package is mature enough to be considered stable.
