### 2021-07-15 vgherard/kgrams#2 @vgherard
Documentation of standards via `srr`
See the [rOpenSci package development guide](https://stats-devguide.ropensci.org/pkgdev.html#pkgdev-srr).

TBD before submission to rOpenSci, c.f. the presubmission inquiry ropensci/software-review#452.
