### 2021-07-15 #452 @noamross
Thank you for our first statistical package pre-submission, @vgherard!  I believe this clearly falls in scope and look forward to a full submission once you have incorporated the `srr` standards component.  I am querying the editorial board to ask for an opinion as to whether this package should also apply standards from the Supervised or Unsupervised learning categories.
