### 2021-07-15 #452 @vgherard
Thanks @noamross, great :) I will begin looking to the `srr` standards, then. It may take me some time, but I'm up for it. Earlier I did a quick check with `autotest` and it seems like there's some trouble in parsing some of my examples, let's see if I can get it to work quickly.
