### 2021-07-16 44f71e2
Change `verbose` defaults to `FALSE`
