### 2021-07-16 #452 @vgherard
Thanks @noamross (@mpadge), I've filed an issue at ropensci-review-tools/autotest#49
