### 2021-07-17 3c23d8f
Set `verbose` default to FALSE
