+++ man/kgram_freqs.Rd
@@ -30,7 +30,7 @@ kgram_freqs(object, ...)
   .tknz_sent = identity,
   dict = NULL,
   open_dict = is.null(dict),
-  verbose = TRUE,
+  verbose = FALSE,
   ...
 )
 
@@ -41,7 +41,7 @@ kgram_freqs(object, ...)
   .tknz_sent = identity,
   dict = NULL,
   open_dict = is.null(dict),
-  verbose = TRUE,
+  verbose = FALSE,
   max_lines = max_lines,
   batch_size = NULL,
   ...
@@ -54,7 +54,7 @@ process_sentences(
   .tknz_sent = attr(freqs, ".tknz_sent"),
   open_dict = TRUE,
   in_place = TRUE,
-  verbose = TRUE,
+  verbose = FALSE,
   ...
 )
 
@@ -65,7 +65,7 @@ process_sentences(
   .tknz_sent = attr(freqs, ".tknz_sent"),
   open_dict = TRUE,
   in_place = TRUE,
-  verbose = TRUE,
+  verbose = FALSE,
   ...
 )
 
@@ -76,7 +76,7 @@ process_sentences(
   .tknz_sent = attr(freqs, ".tknz_sent"),
   open_dict = TRUE,
   in_place = TRUE,
-  verbose = TRUE,
+  verbose = FALSE,
   max_lines = Inf,
   batch_size = max_lines,
   ...
