+++ NAMESPACE
@@ -20,8 +20,8 @@ S3method(parameters,kgram_freqs)
 S3method(parameters,language_model)
 S3method(perplexity,character)
 S3method(perplexity,connection)
-S3method(print,dictionary)
 S3method(print,kgram_freqs)
+S3method(print,kgrams_dictionary)
 S3method(print,kgrams_word_context)
 S3method(print,language_model)
 S3method(probability,character)
@@ -30,11 +30,11 @@ S3method(process_sentences,character)
 S3method(process_sentences,connection)
 S3method(query,kgram_freqs)
 S3method(query,kgrams_dictionary)
-S3method(str,dictionary)
 S3method(str,kgram_freqs)
+S3method(str,kgrams_dictionary)
 S3method(str,language_model)
-S3method(summary,dictionary)
 S3method(summary,kgram_freqs)
+S3method(summary,kgrams_dictionary)
 S3method(summary,language_model)
 export("%+%")
 export("%|%")
