+++ R/srr-stats-standards.R
@@ -1,0 +1,180 @@
+#' srr_stats
+#'
+#' All of the following standards initially have `@srrstatsTODO` tags.
+#' These may be moved at any time to any other locations in your code.
+#' Once addressed, please modify the tag from `@srrstatsTODO` to `@srrstats`,
+#' or `@srrstatsNA`, ensuring that references to every one of the following
+#' standards remain somewhere within your code.
+#' (These comments may be deleted at any time.)
+#'
+#' @srrstatsVerbose TRUE
+#'
+#' @srrstatsTODO {G1.0} *Statistical Software should list at least one primary reference from published academic literature.* 
+#' @srrstatsTODO {G1.1} *Statistical Software should document whether the algorithm(s) it implements are:* - *The first implementation of a novel algorithm*; or - *The first implementation within **R** of an algorithm which has previously been implemented in other languages or contexts*; or - *An improvement on other implementations of similar algorithms in **R***. 
+#' @srrstatsTODO {G1.2} *Statistical Software should include a* Life Cycle Statement *describing current and anticipated future states of development.* 
+#' @srrstatsTODO {G1.3} *All statistical terminology should be clarified and unambiguously defined.* 
+#' @srrstatsTODO {G1.4} *Software should use [`roxygen2`](https://roxygen2.r-lib.org/) to document all functions.*
+#' @srrstatsTODO {G1.4a} *All internal (non-exported) functions should also be documented in standard [`roxygen2`](https://roxygen2.r-lib.org/) format, along with a final `@noRd` tag to suppress automatic generation of `.Rd` files.* 
+#' @srrstatsTODO {G1.5} *Software should include all code necessary to reproduce results which form the basis of performance claims made in associated publications.* 
+#' @srrstatsTODO {G1.6} *Software should include code necessary to compare performance claims with alternative implementations in other R packages.* 
+#' @srrstatsTODO {G2.0} *Implement assertions on lengths of inputs, particularly through asserting that inputs expected to be single- or multi-valued are indeed so.*
+#' @srrstatsTODO {G2.0a} Provide explicit secondary documentation of any expectations on lengths of inputs
+#' @srrstatsTODO {G2.1} *Implement assertions on types of inputs (see the initial point on nomenclature above).*
+#' @srrstatsTODO {G2.1a} *Provide explicit secondary documentation of expectations on data types of all vector inputs.*
+#' @srrstatsTODO {G2.2} *Appropriately prohibit or restrict submission of multivariate input to parameters expected to be univariate.*
+#' @srrstatsTODO {G2.3} *For univariate character input:*
+#' @srrstatsTODO {G2.3a} *Use `match.arg()` or equivalent where applicable to only permit expected values.*
+#' @srrstatsTODO {G2.3b} *Either: use `tolower()` or equivalent to ensure input of character parameters is not case dependent; or explicitly document that parameters are strictly case-sensitive.*
+#' @srrstatsTODO {G2.4} *Provide appropriate mechanisms to convert between different data types, potentially including:*
+#' @srrstatsTODO {G2.4a} *explicit conversion to `integer` via `as.integer()`*
+#' @srrstatsTODO {G2.4b} *explicit conversion to continuous via `as.numeric()`*
+#' @srrstatsTODO {G2.4c} *explicit conversion to character via `as.character()` (and not `paste` or `paste0`)*
+#' @srrstatsTODO {G2.4d} *explicit conversion to factor via `as.factor()`*
+#' @srrstatsTODO {G2.4e} *explicit conversion from factor via `as...()` functions*
[…]
