+++ DESCRIPTION
@@ -14,7 +14,7 @@ Description:
 License: GPL (>= 3)
 Encoding: UTF-8
 LazyData: true
-Roxygen: list(markdown = TRUE)
+Roxygen: list(markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))
 RoxygenNote: 7.1.1
 SystemRequirements: C++11
 LinkingTo: 
