### 2021-07-23 fcf6ba2
Preliminary categorization of srrstats standards
