+++ R/perplexity.R
@@ -5,6 +5,9 @@
 #' @author Valerio Gherardi
 #' @md
 #'
+#' @srrstats {G2.0a} 
+#' *Provide explicit secondary documentation of any expectations on lengths of inputs*
+#'
 #' @param text a character vector or connection. Test corpus from which 
 #' language model perplexity is computed.
 #' @param model either an object of class \code{language_model}, or a 
@@ -72,6 +75,10 @@
 #' sapply(c("N = 4" = 4, "N = 6" = 6, "N = 8" = 8), FUN)
 #' }
 #' 
+#' @srrstats {G1.0} 
+#' *Statistical Software should list at least one primary reference from published academic literature.*
+#' This documentation page provides references for the definition of the perplexity metric.
+#' 
 #' @references 
 #' \insertAllCited{}
 #' 
@@ -120,7 +127,6 @@ perplexity.connection <- function(
         ...
 ) 
 {
-        
         if (is.null(batch_size)) 
                 batch_size <- -1L
         
