### 2021-07-26 9a58255
Add assertions.R for argument checking
