### 2021-07-26 5a42b84
remove srr stats tags
