### 2021-07-26 4a5cd0e
Use assertions in new_kgram_freqs()
