### 2021-07-26 1b5297d
Add tests for 'dictionary'
