### 2021-07-26 7d76af8
Format message in assert_true_or_false
