### 2021-07-26 cbee9ca
as_dictionary(NULL) now returns an empty dictionary
