### 2021-07-26 f83337b
Add NEWS.md
