+++ NEWS.md
@@ -1,0 +1,3 @@
+# kgrams 0.1.0.9000
+
+* Added a `NEWS.md` file to track changes to the package.
