### 2021-07-26 9c43ab4
Assertions for S3 class kgram_freqs
