### 2021-07-26 8554686
Test process_sentences_init()
