### 2021-07-26 8492cb8
Use standard method dispatch for calling the right process_sentences() method
