### 2021-07-26 ecbf31e
Add assertions before method dispatch in process_sentences()
