### 2021-07-26 07f516f
Fix bug in kgram_process_task
