+++ NEWS.md
@@ -1,7 +1,11 @@
 # kgrams (development version)
 
-### Software Improvements
+### Improvements
 * Extended testing and improved argument checking
+
+### Bug Fixes
+* Fixed bug which caused .preprocess and .tknz_sent arguments to be impossible 
+to replace with `process_sentences()`.
 
 ### New features
 * `as_dictionary(NULL)` now returns an empty `dictionary`.
