### 2021-07-26 dfb4b8c
Test kgram_process_task
