### 2021-07-26 13a656e
Add unit tests for kgram_freqs and process_sentences
