### 2021-07-26 80ce9de
Modify wrong defaults for 'max_lines' and 'batch_size' in kgram_freqs.connection()
