### 2021-07-26 e4f48d8
Mention fixes for 'max_lines' and 'batch_size' defaults
