### 2021-07-26 03c94ce
add 'can_be_inf' argument to 'assert_positive_integer()'
