### 2021-07-26 983078f
Add assertions for batch_size and max_lines arguments
