### 2021-07-26 fee9fd6
Test the existence of print methods
