### 2021-07-26 c6ffcbe
Add assert_probability()
