+++ R/assertions.R
@@ -1,15 +1,11 @@
 #' @srrstats {G2.0} 
-#' *Implement assertions on lengths of inputs.* 
+#' *Implement assertions on lengths of inputs*
 
 kgrams_domain_error <- function(name, what) {
         h <- "Invalid argument"
         x <- paste0("'", name, "'", " must be ", what, ".")
         rlang::abort(c(h, x = x), class = "kgrams_domain_error")
 }
-
-is_number <- function(x)
-        is.numeric(x) && length(x) == 1 && !is.na(x)
-
 
 assert_positive_integer <- function(x, can_be_inf = FALSE) {
         
@@ -24,6 +20,16 @@ assert_positive_integer <- function(x, can_be_inf = FALSE) {
                 name = deparse(substitute(x)), 
                 what = "a length one positive integer"
                 )
+}
+
+assert_probability <- function(x) {
+        p <- is.numeric(x) && length(x) == 1 && !is.na(x) && 0 <= x && x <= 1
+        if (p)
+                return(invisible(NULL))
+        kgrams_domain_error(
+                name = deparse(substitute(x)), 
+                what = "a number between 0 and 1"
+        )
 }
 
 assert_function <- function(x) {
