### 2021-07-26 2b6e27f
Big restructuring of S3 dictionary class
