+++ R/dictionary.R
@@ -1,4 +1,4 @@
-new_dictionary <- function(cpp_obj) {
+new_dictionary <- function(cpp_obj = new(Dictionary)) {
         structure(list(), cpp_obj = cpp_obj, class = "kgrams_dictionary")
 }
 
@@ -27,9 +27,8 @@ str.kgrams_dictionary <- function(object, ...) summary(object)
 #' @md
 #' @srrstats {G2.0a} *Provide explicit secondary documentation of any expectations on lengths of inputs*
 #'
-#' @param text a character vector, a connection or missing/\code{NULL}. 
-#' Source of text from which k-gram frequencies are to be extracted.
-#' @param object an object to be coerced to dictionary.
+#' @param object object from which to extract a dictionary, or to be coerced to 
+#' dictionary.
 #' @param .preprocess a function taking a character vector as input and returning
 #' a character vector as output. Optional preprocessing transformation  
 #' applied to text before creating the dictionary.
@@ -39,31 +38,30 @@ str.kgrams_dictionary <- function(object, ...) summary(object)
 #' Predefined text coverage fraction of the dictionary 
 #' (the most frequent words providing the required coverage are retained).
 #' @param thresh either \code{NULL} or length one a positive integer. 
-#' Predefined text coverage fraction of the dictionary 
-#' (the most frequent words providing the required coverage are retained).
-#' @param batch_size a length one positive integer or \code{NULL}.
-#' Size of text batches when reading text from a \code{file} or a generic 
-#' \code{connection}. If \code{NULL}, all input text is processed in a single 
-#' batch.
+#' Minimum word count threshold to include a word in the dictionary.
+#' @param max_lines a length one positive integer or \code{Inf}.
+#' Maximum number of lines to be read from the \code{connection}. 
+#' If \code{Inf}, keeps reading until the End-Of-File.
+#' @param batch_size a length one positive integer less than or equal to
+#' \code{max_lines}.Size of text batches when reading text from 
+#' \code{connection}.
 #' @param ... further arguments passed to or from other methods.
 #' @param x a \code{dictionary}.
 #' 
 #' @return A \code{dictionary} for \code{dictionary()} and 
 #' \code{as_dictionary()}, a character vector for the \code{as.character()}
 #' method.
-#' @details These generic functions are used to build dictionaries from a text
-#' source, or to coerce other formats to \code{dictionary}, and from a 
+#' @details These generic functions are used to build \code{dictionary} objects, 
+#' or to coerce from other formats to \code{dictionary}, and from a 
 #' \code{dictionary} to a character vector. By now, the only 
 #' non-trivial type coercible to \code{dictionary} is \code{character}, 
 #' in which case each entry of the input vector is considered as a single word.
 #' Coercion from \code{dictionary} to \code{character} returns the list of
 #' words included in the dictionary as a regular character vector.
 #' 
-#' Dictionaries can be \emph{built} from text coming either directly from a 
-#' character vector, or from a connection. The second option is useful if one 
-#' wants to avoid loading the full text corpus in physical memory, 
-#' allowing to process text from different sources such as files, compressed 
-#' files or URLs.
+#' Dictionaries can be extracted from \code{kgram_freqs} objects, or \emph{built} 
+#' from text coming either directly from a character vector or a connection to 
+#' textual input.
 #' 
 #' A single preprocessing transformation can be applied before processing the 
 #' text for unique words. After preprocessing, 
@@ -100,84 +98,96 @@ NULL
 
 #' @rdname dictionary
 #' @export
-dictionary <- function(text, ...) {
-        if (missing(text) || is.null(text))
-                return(dictionary_missing())
[…]
