+++ NAMESPACE
@@ -7,6 +7,7 @@ S3method(as_dictionary,character)
 S3method(as_dictionary,kgrams_dictionary)
 S3method(dictionary,character)
 S3method(dictionary,connection)
+S3method(dictionary,kgram_freqs)
 S3method(kgram_freqs,character)
 S3method(kgram_freqs,connection)
 S3method(kgram_freqs,kgram_freqs)
