### 2021-07-26 ab56780
Replace test on creation of dictionary from char
