### 2021-07-26 7d01fda
Testing of size/cov/thresh arguments
