### 2021-07-26 f8b00a5
skip tests depending on format(methods()) for R < 4
