### 2021-07-26 66ec647
Remove AppVeyor and CircleCI
