### 2021-07-26 411a513
Remove development mode auto
