+++ _pkgdown.yml
@@ -1,6 +1,4 @@
 url: https://vgherard.github.io/kgrams/, https://github.com/vgherard/kgrams
-development:
-  mode: auto
 
 home:
 
