+++ _pkgdown.yml
@@ -1,7 +1,6 @@
 url: https://vgherard.github.io/kgrams/, https://github.com/vgherard/kgrams
 
 home:
-
   links:
   - text: Hands-on introduction to k-gram language models
     href: https://datascienceplus.com/an-introduction-to-k-gram-language-models-in-r/
