+++ pkgdown.yml
@@ -1,0 +1,10 @@
+pandoc: 2.7.3
+pkgdown: 1.6.1
+pkgdown_sha: ~
+articles:
+  kgrams: kgrams.html
+last_built: 2021-07-26T17:56Z
+urls:
+  reference: https://vgherard.github.io/kgrams/, https://github.com/vgherard/kgrams/reference
+  article: https://vgherard.github.io/kgrams/, https://github.com/vgherard/kgrams/articles
+
