### 2021-07-27 0bfa92b
Change 'smoother' to 'language_model' attribute
