### 2021-07-27 e1452ab
Assertions for 'smoother' argument
