### 2021-07-27 6c80b46
Test smoothers() returns a char
