+++ tests/testthat/test-smoothers.R
@@ -1,0 +1,3 @@
+test_that("smoothers() returns a character vector", {
+        expect_vector(smoothers(), character())
+})
