### 2021-07-27 78abf10
Test assert_smoother
