### 2021-07-27 285ea2d
Test validate_smoother
