+++ tests/testthat/test-validate_smoother.R
@@ -11,3 +11,35 @@ test_that("list_parameters has the required structure", {
                 }
         }
 })
+
+test_that("smoother_par_missing throws a warning of the correct class", {
+        expect_warning(
+                smoother_par_missing("smoother", "par", "default"),
+                class = "kgrams_smoother_par_missing_warning"
+                )
+})
+
+test_that("smoother_par_missing throws a warning only once", {
+        capture_output(smoother_par_missing("smoother", "par", 1))
+        expect_warning(
+                smoother_par_missing("smoother", "par", 1),
+                NA
+        )
+})
+
+test_that("smoother_par_error throws an error of the correct class", {
+        expect_error(
+                smoother_par_error("smoother", "par", "expected"),
+                class = "kgrams_smoother_par_error"
+        )
+})
+
+test_that("validate_smoother does not throw errors on default values", {
+        for (smoother in smoothers()) {
+                l <- list_parameters(smoother)
+                args <- lapply(l, function(x) x$default)
+                names(args) <- sapply(l, function(x) x$name)
+                args <- c(list(smoother = smoother), args)
+                expect_error(do.call(validate_smoother, args), NA)
+        }
+})
