### 2021-07-27 fdcbaf2
Use assertion for argchecking
