### 2021-07-27 bb35429
Test a warning and an error from validate_smoother
