### 2021-07-27 0b78090
Use assertion on 'N' input
