### 2021-07-27 a0dbb3a
Test 'language_model' print methods
