### 2021-07-27 4ae3c67
Test assert_string
