+++ tests/testthat/test-assertions.R
@@ -12,6 +12,15 @@ test_that("assert_positive_integer", {
         expect_error(assert_positive_integer(Inf, can_be_inf = TRUE), NA)
 })
 
+test_that("assert_string", {
+        expect_error(assert_string("a string"), NA)
+        
+        class <- "kgrams_domain_error"
+        expect_error(assert_string(1), class = class)
+        expect_error(assert_string(c("a", "b")), class = class)
+        expect_error(assert_string(NA_character_), class = class)
+})
+
 test_that("assert_probability", {
         class <- "kgrams_domain_error"
         expect_error(assert_probability(0.5), NA)
