+++ R/assertions.R
@@ -7,24 +7,35 @@ kgrams_domain_error <- function(name, what) {
         rlang::abort(c(h, x = x), class = "kgrams_domain_error")
 }
 
+assert_number <- function(x, name = deparse(substitute(x)))
+{
+        if (is.numeric(x) && length(x) == 1 && !is.na(x))
+                return(invisible(NULL))
+        kgrams_domain_error(name = name, what = "a length one numeric (not NA)")
+}
+
+assert_string <- function(x, name = deparse(substitute(x))) {
+        if (is.character(x) && length(x) == 1 && !is.na(x))
+                return(invisible(NULL))
+        kgrams_domain_error(name, what = "a length one character (not NA)")
+}
+
 assert_positive_integer <- function(
         x, can_be_inf = FALSE, name = deparse(substitute(x))
         ) 
 {
-        p <- is.numeric(x) && length(x) == 1 && !is.na(x) &&
-                (
-                        (is.infinite(x) && can_be_inf) || 
-                        (!is.infinite(x) && as.integer(x) == x && x > 0)
-                )
+        assert_number(x, name = name)
+        p <- (is.infinite(x) && can_be_inf) || 
+                (!is.infinite(x) && as.integer(x) == x && x > 0)
         if (p) 
                 return(invisible(NULL))
-        kgrams_domain_error(name = name, what = "a length one positive integer")
+        kgrams_domain_error(name = name, what = "a positive integer")
 }
 
 assert_probability <- function(x, name = deparse(substitute(x))) 
 {
-        p <- is.numeric(x) && length(x) == 1 && !is.na(x) && 0 <= x && x <= 1
-        if (p)
+        assert_number(x, name = name)
+        if (0 <= x && x <= 1)
                 return(invisible(NULL))
         kgrams_domain_error(name = name, what = "a number between 0 and 1")
 }
@@ -36,8 +47,7 @@ assert_function <- function(x, name = deparse(substitute(x))) {
 }
 
 assert_true_or_false <- function(x, name = deparse(substitute(x))) {
-        p <- is.logical(x) && length(x) == 1 && !is.na(x)
-        if (p)
+        if (is.logical(x) && length(x) == 1 && !is.na(x))
                 return(invisible(NULL))
         kgrams_domain_error(name = name, what = "TRUE or FALSE")
 }
@@ -56,8 +66,7 @@ assert_kgram_freqs <- function(x, name = deparse(substitute(x))) {
 }
 
 assert_smoother <- function(x, name = deparse(substitute(x))) {
-        if (!is.character(x) || length(x) != 1 || is.na(x))
-                kgrams_domain_error(name, "a length one character (not NA).")
+        assert_string(x, name = name)
         if (!(x %in% smoothers()))
                 rlang::abort(
                         message = c("Invalid smoother",
@@ -67,4 +76,3 @@ assert_smoother <- function(x, name = deparse(substitute(x))) {
                         class = c("kgrams_smoother_error")
                 )
 }
-        
