### 2021-07-27 9f81073
Use assert_string
