+++ R/parameters.R
@@ -79,6 +79,7 @@ param <- function(object, which) {
 
 #' @export
 param.language_model <- function(object, which) {
+        validate_parameter(object, which)
         res <- attr(object, "cpp_obj")[[which]]
         return(res)
 }
@@ -86,6 +87,7 @@ param.language_model <- function(object, which) {
 #' @rdname parameters
 #' @export
 param.kgram_freqs <- function(object, which) {
+        validate_parameter(object, which)
         res <- attr(object, "cpp_obj")[[which]]
         return(res)
 }
@@ -103,21 +105,7 @@ param.kgram_freqs <- function(object, which) {
 
 #' @export
 `param<-.language_model` <- function(object, which, value) {
-        
-        if (which == "V") {
-                h <- paste0("Dictionary size cannot be set.")
-                rlang::abort(h, class = "read_only_par_error")
-        }
-        
-        if ( is.null(attr(object, "cpp_obj")[[which]]) ) {
-                smoother <- attr(object, "smoother")
-                h <- paste0("\"", which, 
-                            "\" is not a valid parameter for smoother \"",
-                            smoother, "\".")
-                i <- paste0("See info(\"", smoother, "\") or ?param.")
-                rlang::abort(c(h, i = i), class = "smoother_invalid_parameter")
-        }
-        
+        validate_parameter(object, which, value)
         attr(object, "cpp_obj")[[which]] <- value
         return(object)
 }
@@ -139,4 +127,58 @@ parameters.language_model <- function(object) {
         res <- lapply(names, function(name) param(object, name))
         names(res) <- names
         return(res)
+}
+
+#---------------------------------------------------------- Parameter validation
+
+validate_parameter <- function(object, parameter, value) 
+        UseMethod("validate_parameter")
+
+validate_parameter.language_model <- function(object, which, value) {
+        smoother <- attr(object, "smoother")
+        valid_params <- sapply(list_parameters(smoother), function(x) x$name)
+        valid_params <- c("N", "V", valid_params)
+        
+        if (!(which %in% valid_params)) {
+                h <- "Unknown parameter"
+                x <- paste0("'", which, "' is not a recognized parameter")
+                i <- paste0("See info(\"", smoother, "\") or ?param.")
+                msg <- c(h,x = x, i = i)
+                rlang::abort(msg, class = "kgrams_unknown_par_error")
+        }
+        
+        if (missing(value))
+                return(invisible(NULL))
+        
+        if (which == "V") {
+                h <- "Read only parameter" 
+                x <- "Dictionary size cannot be set."
+                rlang::abort(c(h, x = x), class = "kgrams_read_only_par_error")
+        }
+        
+        if (which == "N" && value > attr(object, "cpp_freqs")$N) {
+                h <- "Invalid parameter" 
+                x <- paste0("'N' cannot be larger than the order ",
+                            "of the underlying k-gram frequency table (",
+                            attr(object, "cpp_freqs")$N, ").")
+                rlang::abort(c(h, x = x), class = "kgrams_invalid_par_error")
+        }
+        
+        l <- list_parameters(smoother)
+        args <- lapply(l, function(x) x$default)
+        names(args) <- sapply(l, function(x) x$name)
+        args <- c(list(smoother = smoother), args)
+        args[["which"]] <- value
+        
+        do.call(validate_smoother, args)
+}
+
+validate_parameter.kgram_freqs <- function(object, which, value) {
+        if (!(which %in% c("N", "V"))) {
+                h <- "Unknown parameter"
+                x <- paste0("'", which, "' is not a recognized parameter.")
+                i <- paste0("See ?param.")
+                msg <- c(h,x = x, i = i)
+                rlang::abort(msg, class = "kgrams_unknown_par_error")
+        }
 }
