### 2021-07-29 46ffa5e
Remove Rcpp:: namespace identifier (using namespace Rcpp)
