### 2021-07-29 aa41fb9
Throw error if input contains NA
