+++ src/UtilitiesR.cpp
@@ -19,7 +19,7 @@ CharacterVector preprocess(
         auto itend = input.end();
         for(auto it = input.begin(); it != itend; ++it) {
                 if (*it == Rcpp::NA)
-                        stop("'input' cannot contain NA values.");
+                        continue;
                 temp = *it;
                 if (erase != "") temp = std::regex_replace(temp, erase_, "");
                 if (lower_case) for (char& c : temp) c = tolower(c);
