### 2021-07-29 5bbea80
Test behaviour with NAs
