### 2021-07-29 68d7a3e
simplify expectation for preprocess(character())
