### 2021-07-30 6060c1f
Throw error on NA input
