### 2021-07-30 d110822
Fix starting position of next sentence
