### 2021-07-30 c4cf46c
Test corner case of "" input
