### 2021-07-30 da629a5
Add tests for tknz_sent
