### 2021-07-30 b965404
Add [ subsetting ops for kgram_freqs and dictionary
