+++ R/query.R
@@ -33,6 +33,9 @@
 #' \code{paste("i love", UNK())}). Queries from k-grams of order \code{k > N}
 #' will return \code{NA}.
 #' 
+#' The subsetting form \code{object[x]} is equivalent to 
+#' \code{query(object, x)}.
+#' 
 #' See also the examples below.
 #'    
 #' @examples
@@ -43,11 +46,14 @@
 #' identical(query(f, "c"), query(f, "d"))  # TRUE, both "c" and "d" are <UNK>
 #' identical(query(f, UNK()), query(f, "c")) # TRUE
 #' query(f, EOS()) # 1, since text is a single sentence
+#' f[c("b b", "b")] # query with subsetting synthax 
 #' 
 #' # Querying a dictionary
 #' d <- as_dictionary(c("a", "b"))
 #' query(d, c("a", "b", "c")) # query some words
-#' query(f, c(BOS(), EOS(), UNK())) # c(TRUE, TRUE, FALSE)
+#' query(d, c(BOS(), EOS(), UNK())) # c(TRUE, TRUE, FALSE)
+#' d["a"] query with subsetting synthax
+#' 
 #' @name query
 
 #' @rdname query
@@ -63,6 +69,15 @@ query.kgram_freqs <- function(object, x) {
 
 #' @rdname query
 #' @export
+`[.kgram_freqs` <- query.kgram_freqs
+        
+
+#' @rdname query
+#' @export
 query.kgrams_dictionary <- function(object, x) {
         attr(object, "cpp_obj")$query(x)
 }
+
+#' @rdname query
+#' @export
+`[.kgrams_dictionary` <- query.kgrams_dictionary
