+++ NAMESPACE
@@ -1,5 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method("[",kgram_freqs)
+S3method("[",kgrams_dictionary)
 S3method("param<-",kgram_freqs)
 S3method("param<-",language_model)
 S3method(as.character,kgrams_dictionary)
