### 2021-07-30 23ea947
Add assert_character_no_NA test
