+++ man/query.Rd
@@ -56,7 +56,9 @@ the word \code{"prcsrn"} is outside the dictionary, querying
 will return \code{NA}.
 
 The subsetting form \code{object[x]} is equivalent to
-\code{query(object, x)}.
+\code{query(object, x)}. The query of the empty string \code{""} returns the
+total count of words, including the \code{EOS} and \code{UNK} tokens, but not
+the \code{BOS} token.
 
 See also the examples below.
 }
@@ -69,6 +71,7 @@ identical(query(f, "c"), query(f, "d"))  # TRUE, both "c" and "d" are <UNK>
 identical(query(f, UNK()), query(f, "c")) # TRUE
 query(f, EOS()) # 1, since text is a single sentence
 f[c("b b", "b")] # query with subsetting synthax 
+f[""] # 9 (includes the EOS token)
 
 # Querying a dictionary
 d <- as_dictionary(c("a", "b"))
