### 2021-07-30 e2bbdaf
Test assert_character_no_NA
