+++ tests/testthat/test-string_concatenation.R
@@ -1,0 +1,3 @@
+test_that("minimal test to ensure correct definition", {
+        expect_identical("a" %+% "b", "a b")
+})
