### 2021-07-30 7ebab0c
Temporary fix for R < 4 test w/ infix op
