### 2021-07-30 5fc084d
Improve code readability
