### 2021-07-30 a2b28bf
assert_character_no_NA for character input
