### 2021-07-30 c74c3d6
Fix smoother check
