### 2021-07-30 bf89e76
Assert language_model in perplexity
