### 2021-07-30 df0484a
Add TODO perplexity() tests
