### 2021-08-02 d7e14b2
Accept only language_model class objects in perplexity()
