### 2021-08-02 772d69f
Test check_model_perplexity
