### 2021-08-02 e2a5d4a
assert_positive_number
