### 2021-08-02 41e79a1
Assertions in sample_sentences()
