### 2021-08-02 749bedd
Add test for sample_sentences()
