### 2021-08-02 c0dfda7
Test errors for more than one specification of cov, thresh, size in dictionary()
