### 2021-08-02 f8c6e9e
Add missing unit tests for dictionary()
