### 2021-08-02 f04776a
Remove unused "laplace" smoother
