### 2021-08-02 6b5c474
Accept only 'language_model' class objects as 'model' argument
