### 2021-08-02 1986338
Remove 'as_language_model' helper
