### 2021-08-02 f2e87af
Remove example computing probabilities from kgram_freqs()
