+++ README.Rmd
@@ -25,11 +25,14 @@ knitr::opts_chunk$set(
 [![Codecov test coverage](https://codecov.io/gh/vgherard/kgrams/branch/main/graph/badge.svg)](https://codecov.io/gh/vgherard/kgrams?branch=main)
 [![CRAN status](https://www.r-pkg.org/badges/version/kgrams)](https://CRAN.R-project.org/package=kgrams)
 [![R-universe status](https://vgherard.r-universe.dev/badges/kgrams)](https://vgherard.r-universe.dev/)
+[![Website](https://img.shields.io/badge/Website-here-blue)](https://vgherard.github.io/kgrams/)
 <!-- badges: end -->
 
-`kgrams` provides tools for training and evaluating $k$-gram language models, including several probability smoothing methods, perplexity computations, random text generation and more. It is based on an C++ backend (which can be used itself as a standalone library for $k$-gram based NLP) which makes `kgrams` fast, coupled with an accessible R API which aims at streamlining the process of model building, and can be suitable for small- and medium-sized NLP experiments, baseline model building, and for pedagogical purposes.
+[`kgrams`](https://vgherard.github.io/kgrams/) provides tools for training and evaluating $k$-gram language models, including several probability smoothing methods, perplexity computations, random text generation and more. It is based on an C++ back-end which makes `kgrams` fast, coupled with an accessible R API which aims at streamlining the process of model building, and can be suitable for small- and medium-sized NLP experiments, baseline model building, and for pedagogical purposes.
 
-**UPDATE** Check out my hands-on [tutorial post on $k$-gram language models](https://datascienceplus.com/an-introduction-to-k-gram-language-models-in-r/) using R at [DataScience+](https://datascienceplus.com/)!
+## For beginners
+If you have no idea about what $k$-gram models are *and* didn't get here by 
+accident, you can check out my hands-on [tutorial post on $k$-gram language models](https://datascienceplus.com/an-introduction-to-k-gram-language-models-in-r/) using R at [DataScience+](https://datascienceplus.com/).
 
 ## Installation
 
@@ -43,16 +46,15 @@ install.packages("kgrams")
 
 #### Development version
 
-You can install the development version from [GitHub](https://github.com/vgherard/kgrams) with:
+You can install the development version from [my R-universe](https://vgherard.r-universe.dev/) with:
 
 ``` r
-# install.packages("devtools")
-devtools::install_github("vgherard/kgrams")
+install.packages("kgrams", repos = "https://vgherard.r-universe.dev/")
 ```
 
 ## Example
 
-This example shows how to train a modified Kneser-Ney 4-gram model on Shakespeare's "Much Ado About Nothing" using `kgrams`.
+This example shows how to train a modified Kneser-Ney 4-gram model on Shakespeare's play "Much Ado About Nothing" using `kgrams`.
 
 ```{r}
 library(kgrams)
@@ -88,10 +90,3 @@ sample_sentences(model = mkn, n = 3, max_length = 10, t = 10)
 ## Getting Help
 
 For further help, you can consult the reference page of the `kgrams` [website](https://vgherard.github.io/kgrams/) or [open an issue](https://github.com/vgherard/kgrams/issues) on the GitHub repository of `kgrams`. A vignette is available on the website, illustrating the process of building language models in-depth.
-
-## Development
-
-This project is in an early developmental stage, thorough tests of the algorithms and unit tests still need to be implemented, many computations leave some room for optimization, the API may change, *etc.*. If you feel like contributing to `kgrams`, here's is some useful information.
-
-Development of `kgrams` takes place on its [GitHub repository](https://github.com/vgherard/kgrams/). If you find a bug,
[…]
