+++ DESCRIPTION
@@ -5,7 +5,7 @@ Authors@R:
     person(given = "Valerio",
            family = "Gherardi",
            role = c("aut", "cre"),
-           email = "vgherard@sissa.it",
+           email = "vgherard840@gmail.com",
            comment = c(ORCID = "0000-0002-8215-3013"))
 Description: 
         Tools for training and evaluating k-gram language models in R, 
