### 2021-11-13 20460c7
Remove `srr` stats standards checks
