+++ DESCRIPTION
@@ -14,7 +14,7 @@ Description:
 License: GPL (>= 3)
 Encoding: UTF-8
 LazyData: true
-Roxygen: list(markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))
+Roxygen: list(markdown = TRUE, roclets = c ("namespace", "rd"))
 RoxygenNote: 7.1.2
 SystemRequirements: C++11
 LinkingTo: 
