### 2021-11-13 e5aaf3e
Remove `srrstats` entries from documentation
