+++ vignettes/kgrams.Rmd
@@ -16,25 +16,6 @@ knitr::opts_chunk$set(
 )
 ```
 
-```{r srr-tags, eval = FALSE, echo = FALSE}
-#' @srrstats {G1.0} 
-#' *Statistical Software should list at least one primary reference from published academic literature.*
-#' Various references are provided in this vignette.
-#'
-#' @srrstats {G1.1} 
-#' *Statistical Software should document whether the algorithm(s) it implements are...*
-#' Clarified in a paragraph of the introduction below.
-#'
-#' @srrstats {G1.3} 
-#' *All statistical terminology should be clarified and unambiguously defined.*
-#' In this vignette, all statistical and probability theory terminology used
-#' by the package and package documentation is carefully defined. Many 
-#' explanations, which should suffice for most users, are also given in regular 
-#' documentation pages; the vignette's purpose is to make everything accessible
-#' also to beginners.
-```
-
-
 ```{r setup}
 library(kgrams)
 ```
