+++ man/smoothers.Rd
@@ -43,6 +43,7 @@ smoothers()
 # Get information on smoother "kn", i.e. Interpolated Kneser-Ney
 info("kn")
 
+
 }
 \references{
 \insertAllCited{}
