+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: kgrams
 Title: Classical k-gram Language Models
-Version: 0.1.1.9000
+Version: 0.1.2
 Authors@R: 
     person(given = "Valerio",
            family = "Gherardi",
