+++ inst/REFERENCES.bib
@@ -38,7 +38,7 @@
     year = "2007",
     address = "Prague, Czech Republic",
     publisher = "Association for Computational Linguistics",
-    url = "https://www.aclweb.org/anthology/D07-1090/",
+    url = "https://aclweb.org/anthology/D07-1090/",
     pages = "858--867",
 }
 
