### 2021-11-13 4b92693
Update CRAN comments
