### 2021-11-13 3b5febc
Correct CRAN link (use canonical form)
