### 2021-11-15 c357195
Reconfigure GitHub CI
