### 2021-11-15 4270783
Update CRAN comments
