### 2021-11-15 a762a21
Install pandoc-citeproc before RCMDCHECK
