+++ .github/workflows/R-CMD-check.yaml
@@ -21,6 +21,11 @@ jobs:
         with:
           use-public-rspm: true
 
+      - name: Install pandoc pandoc-citeproc
+        run: |
+          sudo apt-get update -y
+          sudo apt-get install -y pandoc pandoc-citeproc
+
       - uses: r-lib/actions/setup-r-dependencies@v1
         with:
           extra-packages: rcmdcheck
