### 2022-08-09 #452 @noamross
Hello @vgherard! We're going back to some in-progress submissions that got stuck in an ambiguous state.  Sorry that we haven't reached out in a while.  I just wanted to see if ropensci peer review is something you were still interested in pursuing.
