### 2022-08-11 #452 @vgherard
Dear @noamross thanks for checking in and sorry for the long silence, I totally forgot about this process being open.

Sadly, right now I'm too short of time for a relatively demanding submission like this... Apart from this, over time I became a bit unsatisfied with certain aspects of this package, which I'd at least try to improve before submitting.

I'll close this, with the hope to come back to it in a not too far future :-)

Thanks!
