### 2023-06-06 #452 @vgherard
Dears, thanks for keeping in touch.

I had a look at the requirements I would need to cover in order to submit `{kgrams}`, and again, sorry but this is too much for me.

The output of `pkgcheck()` alone looks intimidating - function names, usage of `<<-`, usage of `sapply()`, etc.etc.. Also, I imagine that passing `autotest` and `srr` would probably be much more demanding.

These are in general quick things, but with a package of the dimension of `{kgrams}` it takes a good amount of effort to finally get the green light - an effort I'm not really interested into, since the only thing I'm doing with that package at the moment is keeping it alive on CRAN :')

It's understood that when I say "too much" I refer only to my individual case - I think the work you're doing by putting up this review process is awesome.

For next package ideas I will definitely consider implementing ropensci standard from the onset!
